+++ apps/ledger-live-mobile/src/screens/Discover/DiscoverCard.tsx
@@ -14,6 +14,7 @@ const DiscoverCard = ({
   disabled,
   cardProps,
   imageContainerProps,
+  subtitleFirst,
   ...props
 }: {
   title: string;
@@ -29,6 +30,7 @@ const DiscoverCard = ({
   testID?: string;
   cardProps?: any;
   imageContainerProps?: any;
+  subtitleFirst?: boolean;
 }) => (
   <Touchable onPress={onPress} disabled={disabled} {...props}>
     <Flex
@@ -51,8 +53,13 @@ const DiscoverCard = ({
         flex={1}
         justifyContent="flex-start"
         alignItems="flex-start"
+        flexDirection={subtitleFirst ? "column-reverse" : "column"}
       >
-        <Flex flexDirection="row" alignItems="center" mb={6}>
+        <Flex
+          flexDirection="row"
+          alignItems="center"
+          mb={subtitleFirst ? 0 : 6}
+        >
           <Text
             mt={2}
             variant={"h2"}
@@ -63,9 +70,9 @@ const DiscoverCard = ({
             {title}
           </Text>
         </Flex>
-
         {subTitle && (
           <Text
+            mb={subtitleFirst ? 2 : 0}
             variant={"body"}
             fontWeight={"medium"}
             color={"neutral.c70"}