+++ apps/ledger-live-mobile/src/screens/Manager/Device/DeviceLanguage.tsx
@@ -1,13 +1,14 @@
-import { Flex, Icons, Text, Button } from "@ledgerhq/native-ui";
+import { Flex, Icons, Text } from "@ledgerhq/native-ui";
 import React, { useCallback, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Language, DeviceInfo } from "@ledgerhq/types-live";
-import { useAvailableLanguagesForDevice } from "@ledgerhq/live-common/lib/manager/hooks";
-import { Device } from "@ledgerhq/live-common/lib/hw/actions/types";
+import { useAvailableLanguagesForDevice } from "@ledgerhq/live-common/manager/hooks";
+import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import BottomModal from "../../../components/BottomModal";
 import DeviceLanguageSelection from "./DeviceLanguageSelection";
 import ChangeDeviceLanguageActionModal from "../../../components/ChangeDeviceLanguageActionModal";
 import { track } from "../../../analytics";
+import Touchable from "../../../components/Touchable";
 
 type Props = {
   pendingInstalls: boolean;
@@ -85,13 +86,21 @@ const DeviceLanguage: React.FC<Props> = ({
           </Text>
         </Flex>
         {availableLanguages.length ? (
-          <Button
-            disabled={pendingInstalls}
-            Icon={Icons.DropdownMedium}
-            onPress={openChangeLanguageModal}
+          <Touchable
+            onPress={pendingInstalls ? undefined : openChangeLanguageModal}
           >
-            {t(`deviceLocalization.languages.${currentDeviceLanguage}`)}
-          </Button>
+            <Flex flexDirection="row" alignItems="center" pr={2}>
+              <Text
+                variant="body"
+                color="primary.c80"
+                fontWeight="semiBold"
+                mr={2}
+              >
+                {t(`deviceLocalization.languages.${currentDeviceLanguage}`)}
+              </Text>
+              <Icons.ChevronRightMedium size={18} color="primary.c80" />
+            </Flex>
+          </Touchable>
         ) : (
           <Text>
             {t(`deviceLocalization.languages.${currentDeviceLanguage}`)}