+++ apps/ledger-live-mobile/src/screens/Onboarding/steps/language.tsx
@@ -13,7 +13,7 @@ import { StackScreenProps } from "@react-navigation/stack";
 import i18next from "i18next";
 import RNRestart from "react-native-restart";
 import { useDispatch, useSelector } from "react-redux";
-import { useAvailableLanguagesForDevice } from "@ledgerhq/live-common/lib/manager/hooks";
+import { useAvailableLanguagesForDevice } from "@ledgerhq/live-common/manager/hooks";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import { from } from "rxjs";
 import { DeviceModelInfo, idsToLanguage, Language } from "@ledgerhq/types-live";