+++ libs/ledgerjs/packages/types-live/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ledgerhq/types-live
 
+## 6.24.3
+
+### Patch Changes
+
+- [#1518](https://github.com/LedgerHQ/ledger-live/pull/1518) [`f255b258a9`](https://github.com/LedgerHQ/ledger-live/commit/f255b258a9c79c5eb89ec65d6647e52b90a9c52c) Thanks [@desirendr](https://github.com/desirendr)! - Types - Add new feature flag FeatureId type
+
+## 6.24.3-hotfix.0
+
+### Patch Changes
+
+- [#1518](https://github.com/LedgerHQ/ledger-live/pull/1518) [`f255b258a9`](https://github.com/LedgerHQ/ledger-live/commit/f255b258a9c79c5eb89ec65d6647e52b90a9c52c) Thanks [@desirendr](https://github.com/desirendr)! - Types - Add new feature flag FeatureId type
+
 ## 6.24.2
 
 ### Patch Changes